ORC-192 Implement zlib compresion stream
Implement zlib compressor based on the output stream. Add corresponding
UTs. Also rename the existing test suite from TestCompression into
TestDecompression.
