+++ c++/src/Compression.hh
@@ -20,8 +20,14 @@
 #define ORC_COMPRESSION_HH
 
 #include "io/InputStream.hh"
+#include "io/OutputStream.hh"
 
 namespace orc {
+
+  enum CompressionStrategy {
+    CompressionStrategy_SPEED = 0,
+    CompressionStrategy_COMPRESSION
+  };
 
   /**
    * Create a decompressor for the given compression kind.
@@ -35,6 +41,23 @@ namespace orc {
                         std::unique_ptr<SeekableInputStream> input,
                         uint64_t bufferSize,
                         MemoryPool& pool);
+
+  /**
+   * Create a compressor for the given compression kind.
+   * @param kind the compression type to implement
+   * @param outStream the output stream that is the underlying target
+   * @param strategy compression strategy
+   * @param bufferCapacity compression stream buffer total capacity
+   * @param blockSize compresssion buffer block size
+   * @param pool the memory pool
+   */
+  std::unique_ptr<BufferedOutputStream>
+     createCompressor(CompressionKind kind,
+                      OutputStream * outStream,
+                      CompressionStrategy strategy,
+                      uint64_t bufferCapacity,
+                      uint64_t blockSize,
+                      MemoryPool& pool);
 }
 
 #endif
